WIP: test eventify
