+++ shed/tests/test_streams.py
@@ -22,7 +22,6 @@ from streamz.core import Stream
 import shed.event_streams as es
 from shed.tests.utils import SinkAssertion
 from ..event_streams import star
-from ..utils import to_event_model
 
 
 def test_clear():
@@ -85,257 +84,6 @@ def test_combine_latest(exp_db, start_uid1, start_uid3):
         assert l1 != l2
         if name == 'event':
             assert l2['seq_num'] == 2
-    for n in ['start', 'descriptor', 'event', 'stop']:
-        assert n in assert_docs
-
-
-def test_eventify(exp_db, start_uid1):
-    source = Stream()
-
-    dp = es.Eventify(source, 'name',
-                     output_info=[('name', {
-                         'dtype': 'str',
-                         'source': 'testing'})])
-    # try two outputs
-    dp2 = es.Eventify(source, 'name', 'name',
-                      output_info=[
-                          ('name', {'dtype': 'str', 'source': 'testing'}),
-                          ('name2', {'dtype': 'str', 'source': 'testing'})])
-
-    # try without output_info, no keys
-    dp3 = es.Eventify(source)
-
-    L = dp.sink_to_list()
-    dp.sink(star(SinkAssertion(False)))
-    dp.sink(print)
-    L2 = dp2.sink_to_list()
-    dp2.sink(star(SinkAssertion(False)))
-    dp2.sink(print)
-
-    L3 = dp3.sink_to_list()
-    dp3.sink(star(SinkAssertion(False)))
-    dp3.sink(print)
-
-    ih1 = exp_db[start_uid1]
-    s = exp_db.restream(ih1, fill=True)
-    for a in s:
-        source.emit(a)
-
-    assert len(L) == 4
-    assert len(L2) == 4
-    assert len(L3) == 4
-    assert_docs = set()
-    assert_docs2 = set()
-    assert_docs3 = set()
-    # zip them since we know they're same length and order
-    for l, l2, l3 in zip(L, L2, L3):
-        assert_docs.add(l[0])
-        assert_docs2.add(l2[0])
-        assert_docs3.add(l3[0])
-        if l[0] == 'event':
-            assert l[1]['data']['name'] == 'test'
-            assert l2[1]['data']['name'] == 'test'
-            assert l2[1]['data']['name2'] == 'test'
-        if l[0] == 'stop':
-            assert l[1]['exit_status'] == 'success'
-            assert l2[1]['exit_status'] == 'success'
-        if l3[0] == 'descriptor':
-            assert l3[1]['data_keys']['name'] == {}
-
-    for n in ['start', 'descriptor', 'event', 'stop']:
-        assert n in assert_docs
-
-
-def test_double_eventify(exp_db, start_uid1):
-    source = Stream()
-
-    dp = es.Eventify(source,
-                     output_info=[('name', {
-                         'dtype': 'str',
-                         'source': 'testing'})])
-    L = dp.sink_to_list()
-    dp.sink(star(SinkAssertion(False)))
-    dp.sink(print)
-
-    ih1 = exp_db[start_uid1]
-    s = exp_db.restream(ih1, fill=True)
-    for a in s:
-        source.emit(a)
-
-    L.clear()
-    for a in exp_db[-1].documents(fill=True):
-        source.emit(a)
-
-    assert len(L) == 4
-    assert_docs = set()
-    for l in L:
-        assert_docs.add(l[0])
-        if l[0] == 'stop':
-            assert l[1]['exit_status'] == 'success'
-    for n in ['start', 'descriptor', 'event', 'stop']:
-        assert n in assert_docs
-
-
-def test_eventify_all(exp_db, start_uid1):
-    source = Stream()
-    # source.sink(print)
-
-    dp = es.Eventify(source)
-    L = dp.sink_to_list()
-    dp.sink(star(SinkAssertion(False)))
-    # dp.sink(print)
-
-    ih1 = exp_db[start_uid1]
-    s = exp_db.restream(ih1)
-    for a in s:
-        source.emit(a)
-
-    assert len(L) == 4
-    assert_docs = set()
-    for l in L:
-        assert_docs.add(l[0])
-        if l[0] == 'event':
-            assert l[1]['data']['name'] == 'test'
-        if l[0] == 'stop':
-            assert l[1]['exit_status'] == 'success'
-    for n in ['start', 'descriptor', 'event', 'stop']:
-        assert n in assert_docs
-
-    L.clear()
-    s = exp_db.restream(ih1)
-    for a in s:
-        source.emit(a)
-
-    assert len(L) == 4
-    assert_docs = set()
-    for l in L:
-        assert_docs.add(l[0])
-        if l[0] == 'event':
-            assert l[1]['data']['name'] == 'test'
-        if l[0] == 'stop':
-            assert l[1]['exit_status'] == 'success'
-    for n in ['start', 'descriptor', 'event', 'stop']:
-        assert n in assert_docs
-
-
-def test_eventify_descriptor(exp_db, start_uid1):
-    source = Stream()
-
-    dp = es.Eventify(source, 'data_keys',
-                     output_info=[('name', {
-                         'dtype': 'str',
-                         'source': 'testing'})],
-                     document='descriptor')
-    # try two outputs
-    dp2 = es.Eventify(source, 'data_keys', 'data_keys',
-                      output_info=[
-                          ('name', {'dtype': 'str', 'source': 'testing'}),
-                          ('name2', {'dtype': 'str', 'source': 'testing'})],
-                      document='descriptor')
-    L = dp.sink_to_list()
-    dp.sink(star(SinkAssertion(False)))
-    dp.sink(print)
-    L2 = dp2.sink_to_list()
-    dp2.sink(star(SinkAssertion(False)))
-    dp2.sink(print)
-
-    ih1 = exp_db[start_uid1]
-    s = exp_db.restream(ih1, fill=True)
-    dk = None
-    for a in s:
-        if a[0] == 'descriptor':
-            print(a)
-            dk = a[1]['data_keys']
-        source.emit(a)
-
-    assert len(L) == 4
-    assert len(L2) == 4
-    assert_docs = set()
-    assert_docs2 = set()
-    # zip them since we know they're same length and order
-    for l, l2 in zip(L, L2):
-        assert_docs.add(l[0])
-        assert_docs2.add(l2[0])
-        if l[0] == 'event':
-            assert l[1]['data']['name'] == dk
-            assert l2[1]['data']['name'] == dk
-            assert l2[1]['data']['name2'] == dk
-        if l[0] == 'stop':
-            assert l[1]['exit_status'] == 'success'
-            assert l2[1]['exit_status'] == 'success'
-    for n in ['start', 'descriptor', 'event', 'stop']:
-        assert n in assert_docs
-
-
-def test_eventify_all_descriptor(exp_db, start_uid1):
-    source = Stream()
-
-    dp = es.Eventify(source, document='descriptor')
-    L = dp.sink_to_list()
-    dp.sink(star(SinkAssertion(False)))
-    dp.sink(print)
-
-    ih1 = exp_db[start_uid1]
-    s = exp_db.restream(ih1, fill=True)
-    dk = None
-    for a in s:
-        if a[0] == 'descriptor':
-            dk = a[1]
-        source.emit(a)
-
-    assert len(L) == 4
-    assert_docs = set()
-    for l in L:
-        assert_docs.add(l[0])
-        if l[0] == 'event':
-            for k in set(dk.keys()) | set(l[1]['data'].keys()):
-                assert l[1]['data'][k] == dk[k]
-        if l[0] == 'stop':
-            assert l[1]['exit_status'] == 'success'
-    for n in ['start', 'descriptor', 'event', 'stop']:
-        assert n in assert_docs
-
-    L.clear()
-    s = exp_db.restream(ih1, fill=True)
-    for a in s:
-        source.emit(a)
-
-    assert len(L) == 4
-    assert_docs = set()
-    for l in L:
-        assert_docs.add(l[0])
-        if l[0] == 'event':
-            for k in set(dk.keys()) | set(l[1]['data'].keys()):
-                assert l[1]['data'][k] == dk[k]
-        if l[0] == 'stop':
-            assert l[1]['exit_status'] == 'success'
-    for n in ['start', 'descriptor', 'event', 'stop']:
-        assert n in assert_docs
-
-
-def test_eventify_no_event():
-    source = Stream()
-
-    dp = es.Eventify(source, 'source',
-                     output_info=[('name', {
-                         'dtype': 'str',
-                         'source': 'testing'})])
-    L = dp.sink_to_list()
-    dp.sink(star(SinkAssertion(False)))
-    dp.sink(print)
-
-    for a in to_event_model([], output_info=[('img', {})]):
-        source.emit(a)
-
-    assert len(L) == 4
-    assert_docs = set()
-    # zip them since we know they're same length and order
-    for l in L:
-        assert_docs.add(l[0])
-        if l[0] == 'event':
-            assert l[1]['data']['name'] == 'to_event_model'
-        if l[0] == 'stop':
-            assert l[1]['exit_status'] == 'success'
     for n in ['start', 'descriptor', 'event', 'stop']:
         assert n in assert_docs
